refactor to 1 function 1 stack for lambda-cfn v2.0.0
