+++ .travis.yml
@@ -2,3 +2,4 @@ language: node_js
 sudo: false
 node_js:
 - 4.3.2
+- 6.10.2
