+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## Unreleased
+
 ## 0.1.2
 
 * Fix lambda function syntax
