--- index.js
@@ -1,2 +1,0 @@
-var lambdaCfn = require('lambda-cfn');
-lambdaCfn.load(module);
