+++ madePublic/function.template.js
@@ -1,0 +1,11 @@
+var lambdaCfn = require('@mapbox/lambda-cfn');
+
+module.exports = lambdaCfn.build({
+  name: 'madePublic',
+  eventSources: {
+    webhook: {
+      method: 'POST',
+      apiKey: false
+    }
+  }
+});
