+++ package.json
@@ -1,7 +1,7 @@
 {
   "license": "BSD-2-Clause",
   "engines": {
-    "node": "4.3.2"
+    "node": "6.10.2"
   },
   "name": "patrol-rules-github",
   "repository": {
@@ -15,20 +15,20 @@
   "version": "0.1.2",
   "dependencies": {
     "github": "0.2.4",
-    "d3-queue": "2.0.3",
-    "lambda-cfn": "^1.0.0"
+    "d3-queue": "^3.0.0",
+    "@mapbox/lambda-cfn": "^2.0.0"
   },
   "scripts": {
-    "test": "NODE_ENV=test tape test/*.test.js test/**/*.test.js",
-    "pretest": "eslint index.js"
+    "test": "eslint **/*.js && NODE_ENV=test tape test/*.test.js test/**/*.test.js",
+    "lint": "eslint **/*.js",
+    "unit-test": "NODE_ENV=test tape test/*.test.js test/**/*.test.js"
   },
   "devDependencies": {
-    "nock": "^7.5.0",
-    "tape": "^4.5.1",
-    "eslint": "^2.4.0",
+    "nock": "^9.0.0",
+    "tape": "^4.0.0",
+    "eslint": "^4.0.0",
     "eslint-config-unstyled": "^1.1.0"
   },
-  "main": "index.js",
   "homepage": "https://github.com/mapbox/patrol-rules-github",
   "description": "GitHub Patrol rules"
 }
