--- test/package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-      "test": "0.0.0"
-  }
-}
